requirements.txt
PyPDF2==3.0.1
pdfminer.six==20231228
pyttsx3==2.90
gTTS==2.5.3
Pillow==10.4.0
moviepy==1.0.3
numpy==1.26.4
